requirements.txt
flask>=3.0.0
requests>=2.31.0
edge-tts>=6.1.9
Pillow>=10.0.0
google-generativeai>=0.8.0
gunicorn>=22.0.0
